+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.
  * Add `CreditCard` constraint.
+ * Add `Ip` constraint.
 
 * 1.0.3 (2015-06-02)