+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.
  * Add `CreditCard` constraint.
  * Add `Ip` constraint.
+ * Add `Iban` constraint.
 
 * 1.0.3 (2015-06-02)