+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+* 2.0.0 (2015-??-??)
+
+ * Make Symfony 2.7 LTS as a minimum requirement.
+
 * 1.1.0 (2015-??-??)
 
  * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.